requirements.txt
# Cosmic Market Oracle - Consolidated Dependencies
# Python Version: 3.10
# Last Updated: 2025-05-19

# =====================================================================
# API Framework
# =====================================================================
fastapi==0.95.0
uvicorn==0.21.0
gunicorn==20.1.0  # Production WSGI server
pydantic==1.10.7
starlette==0.26.1
python-multipart==0.0.6
email-validator==2.0.0

# =====================================================================
# Data Processing & Analysis
# =====================================================================
pandas==2.0.3  # Updated to latest stable
numpy==1.24.3  # Updated to latest stable
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0  # Progress bars

# =====================================================================
# Machine Learning & Deep Learning
# =====================================================================
torch==2.0.1  # Updated to latest stable
torchvision==0.15.1
xgboost==1.7.4
scikit-learn==1.3.0  # Updated to latest stable
lightgbm==3.3.5
catboost==1.1.1

# =====================================================================
# Hyperparameter Optimization
# =====================================================================
optuna==3.3.0  # Updated to latest stable

# =====================================================================
# Database & Storage
# =====================================================================
sqlalchemy==2.0.7
psycopg2-binary==2.9.5  # PostgreSQL adapter for TimescaleDB
psycopg==3.1.8  # Modern PostgreSQL adapter
alembic==1.10.3  # Database migrations
timescaledb-python==0.3.0  # TimescaleDB specific functionality

# =====================================================================
# Astrological Calculations
# =====================================================================
pyswisseph==2.10.3  # Swiss Ephemeris Python bindings
jplephem==2.18  # JPL Ephemeris
skyfield==1.45  # High-precision astronomy

# =====================================================================
# Workflow Orchestration & MLOps
# =====================================================================
apache-airflow==2.5.3
apache-airflow-providers-postgres==5.4.0
mlflow==2.8.0  # Updated to latest stable

# =====================================================================
# Security
# =====================================================================
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
bcrypt==4.0.1
httpx==0.23.3  # HTTP client

# =====================================================================
# API Documentation & GraphQL
# =====================================================================
strawberry-graphql==0.171.1  # GraphQL support

# =====================================================================
# Testing
# =====================================================================
pytest==7.3.1
pytest-cov==4.1.0

# =====================================================================
# Development Tools
# =====================================================================
black==23.3.0  # Code formatting
isort==5.12.0  # Import sorting
flake8==6.0.0  # Linting
mypy==1.2.0  # Type checking

# =====================================================================
# Monitoring & Logging
# =====================================================================
prometheus-client==0.16.0
python-json-logger==2.0.7
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-exporter-otlp==1.15.0

# =====================================================================
# Utilities
# =====================================================================
python-dotenv==1.0.0  # Environment variable management
